Support feet z-axis in meter survey
Summary:
Partially support display of z values in feet for meter surveys
Only 3D scene annotation and mouse readings

Test Plan:
In a meter survey, changed display depths to feet. 3D scene shows Z in feet.
Statusbar and tree show Z in feet.
Position toolbar not yet. 2D viewers not yet.

Reviewers: nanne

Reviewed By: nanne

Subscribers: arnaud

Differential Revision: https://phabricator.dgbes.com/D17741

+++ include/visBase/visannot.h
@@ -70,7 +70,7 @@ public:
 			     axis scale of survey box instead */
     const TrcKeyZSampling& getScale() const;
 
-    void		setScaleFactor(int dim,int scale);
+    void		setScaleFactor(int dim,double scale);
 
     void		setText(int dim,const uiString&);
     void		fillPar(IOPar&) const override;
@@ -95,7 +95,7 @@ private:
     void		getAxisCoords(int,osg::Vec3f&,osg::Vec3f&) const;
     void		setCorner( int, float, float, float );
 
-    int				scalefactor_[3];
+    double		scalefactor_[3];
 
     TrcKeyZSampling		tkzs_;
     TrcKeyZSampling		tkzsdefaultscale_;
